contrib/python/podman/requirements.txt
varlink>=26.1.0
setuptools>=39.2.0
python-dateutil>=2.7.3
